requirements.txt
discord.py==2.3.1
# No need for aioredis lib anymore since redis.py natively supports async coroutines now
redis==4.6.0
asyncpg==0.27.0
# dateparser is required for entering human-readable dates for slash command /whojoined
# Eg.: /whojoined [Discord prompt:since] 30/06/2023 MSK [Discord prompt: until] 01/07/2023 MSK
dateparser==1.1.8
click==8.1.3
sqlalchemy==2.0.17